+++ instrumentation/runner/src/main/kotlin/de/mannodermaus/junit5/AndroidJUnitFrameworkBuilder.kt
@@ -48,13 +48,12 @@ public open class AndroidJUnitFrameworkBuilder internal constructor() : RunnerBu
     }
 
     // One-time parsing setup for runner params, taken from instrumentation arguments
-    private val params by lazy {
+    private val params =
         JUnitFrameworkRunnerParams.create().also { params ->
             // Apply all environment variables & system properties to the running process
             params.registerEnvironmentVariables()
             params.registerSystemProperties()
         }
-    }
 
     @Throws(Throwable::class)
     override fun runnerForClass(testClass: Class<*>): Runner? {
@@ -63,7 +62,7 @@ public open class AndroidJUnitFrameworkBuilder internal constructor() : RunnerBu
 
         try {
             return if (junitFrameworkAvailable) {
-                tryCreateJUnitFrameworkRunner(testClass) { params }
+                tryCreateJUnitFrameworkRunner(testClass, params)
             } else {
                 null
             }